k1c: Multiring port
Summary:
In order to be able to start a system with the correct rights, delegate
everything that is not needed by the debugger to the PL1. In order to be
regular with other layers, we directly jump to PL1 after starting to let
the debugger in PL0 if needed.

Test Plan: Run barebox in integration with latest multiring branch

Reviewers: O73 barebox coolidge, gthouvenin

Reviewed By: O73 barebox coolidge, gthouvenin

Subscribers: gthouvenin

Differential Revision: https://phab.kalray.eu/D432