kvx: exclude dtb from malloc zone
Summary:
When the dtb is inside the malloc zone, when unflattening it, some
malloc will be done. Potentially, these allocation will take place on
the dtb blob and corrupt it. To avoid that, exclude the dtb from the
malloc zone by checking if it overlaps it. If so, determine the largest
zone for the allocation and modify the memory area to use that.

Ref T12629

Test Plan:
Run with JTAG and put the dtb at a specific address right in the beginning
of the malloc zone.

Reviewers: lmorhet, O73 Barebox Coolidge!

Differential Revision: https://phab.kalray.eu/D7186