coop_boot: add kvx-specific commands for secure boot
Summary:
The Secure Boot has a cooperative boot management that implies that all
stages have a way to :
- Retrieve the current boot path
- invalidate a path when it is bad

The last stage of a path has to validate when everything is OK.

This patch adds a command to set a variable to the boot priority of a
path, thus enabling the use of the Barebox bootchooser to pursue correct
operation, and a command to invalidate a path in flash when barebox
detects a corrupted kernel.

Ref T16570

An example is provided as an init script in a defaultenv, and the boot
scripts handle both booting the targets and invalidating current path.

Test Plan: Validation of the whole secure bootchain with RAUC in Linux

Reviewers: O73 Barebox Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D13644