boot: make path invalidation reset the chip
Summary:
While we do not have proper watchdog handling, we cannot simply
invalidate the current path and wait when something goes wrong.
Instead, we have to invalidate the current path and then trig a soft
reset in order to reboot on the other partition.

Ref T17023

Test Plan: Validation of the full-path reliable update in Secure context

Reviewers: O73 Barebox Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T17023

Differential Revision: https://phab.kalray.eu/D15376