common: elf: add size support when reading from a binary
Summary:
Since we will load elf directly from memory, we must ensure we do not go out
of bound when copying segments. For file based loading, the file size is
already check while reading from it so there is no need for size check
in this path.

Test Plan: Wait for other patches to test it

Reviewers: O73 Barebox Coolidge, lmorhet

Reviewed By: O73 Barebox Coolidge, lmorhet

Subscribers: jmaselbas, lmorhet, #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D9453