watchdog: kvx: do not disable watchdog automatically
Summary:
As the watchdog may be set by a lower stage in the boot chain, and for a
good reason (precisely to ensure reliable update when something
unexpected happens in the boot chain), it is a bad idea to
systematically disable it in the probe of a driver.

We should either have a mechanism to refresh it on a regular basis in
barebox itself and set it to something bigger for Linux, or handle it
transparently and let the previous stage set the global watchdog up to
Linux.

This patch assumes for now the latter, as we do not have support for
the kind of regular tasklet required by a regular refresh.

Ref T17312

Test Plan:
Set watchdog to a not-so-high value in the FSBL, and try to make it trig
during the window to stop auto-boot (to ensure driver probes are done)

Reviewers: O73 Barebox Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T17312

Differential Revision: https://phab.kalray.eu/D15755