defconfig: add timeout support
Summary:
The default environment contains an autoboot script using timeout,
but the timeout command is not built by default.
This stops autoboot from pursuing, and prints an error on the console.

This patch enables the timeout command in the default configuration.

Test Plan: Successfuly autoboot from default config and default env.

Reviewers: O73 Barebox Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D19060