sb_cooperative: increase the priority of NEW path
Summary:
The NEW_IMG flag indicates a path that has been programmed, but not yet
verified by the Secure Boot, and was considered as a low priority boot
path.

This, however, creates hassle in non-secure boot setups by changing the
behaviour of the overall boot chain depending on the presence of the
Secure Boot.

This patch raises the priority of NEW images to be above all others, so
that a NEW_IMG path will be chosen when encountered by all boot stages,
thus removing the discrepancy in secure vs non-secure setups.

Encountering NEW images is not a situation that should be expected in
secure setups anyway, so this patch mostly helps non-secure setups.
As this flag is not being protected by a signature, it is not a key
security element.

Test Plan: Rauc test on non-secure setup

Reviewers: O73 Barebox Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D17234