spi-nor: dw-ospi-nor: Limit spi-max-frequency to half the master clock
Summary:
The clock divisor can only divide by even fractions, if the divisor is
set to 0 the output clock is disabled. It is impossible to divide by 1.
The highest possible frequency is limited to half of the master
reference clock frequency.

ref T19223

Test Plan: flash-util on CV2/HAPS

Reviewers: O73 Barebox Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T19223

Differential Revision: https://phab.kalray.eu/D23792