socinfo: fix mppa ID read from fuses
Summary:
The traceability registers order has been fixed on cv2. This requires a
change in the DeviceTree to change the few registers read here, but also
the "register reversal" that was applied on 64b reads.

Ref T1

Test Plan: usb gadget should expose the right MPPA ID on kv3-2

Reviewers: O73 Barebox Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T1

Differential Revision: https://phab.kalray.eu/D32182